Renamed py model module to models
