+++ servicenow/stackstate_checks/servicenow/__init__.py
@@ -2,7 +2,7 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from .__about__ import __version__
-from .model import State, InstanceInfo
+from .models import State, InstanceInfo
 from .servicenow import ServicenowCheck, json_parse_exception
 
 __all__ = [
